+++ src/thztools/thztools.py
@@ -11,8 +11,8 @@ from numpy.fft import irfft, rfft, rfftfreq
 from numpy.random import default_rng
 from numpy.typing import ArrayLike, NDArray
 from scipy import signal
-from scipy.optimize import OptimizeResult, minimize
-from scipy.optimize import approx_fprime as fprime
+from scipy.linalg import sqrtm
+from scipy.optimize import OptimizeResult, approx_fprime, minimize
 
 NUM_NOISE_PARAMETERS = 3
 NUM_NOISE_DATA_DIMENSIONS = 2
@@ -21,7 +21,7 @@ NUM_NOISE_DATA_DIMENSIONS = 2
 @dataclass
 class GlobalOptions:
     r"""
-    Class for global options.
+    Class for storing global options.
 
     Attributes
     ----------
@@ -30,10 +30,11 @@ class GlobalOptions:
         default, times and frequencies are treated as dimensionless quantities
         that are scaled by the (undetermined) sampling time.
     """
-    sampling_time: float | None
+    sampling_time: float | None = None
 
 
-options = GlobalOptions
+#: Instance of ``GlobalOptions`` that stores global options.
+options = GlobalOptions()
 
 
 def set_option(key: str, value: Any) -> None:
@@ -51,6 +52,24 @@ def set_option(key: str, value: Any) -> None:
     -------
     None
         No return value.
+
+    Notes
+    -----
+    Available options, with descriptions:
+
+    sampling_time : float | None
+        Global sampling time, normally in picoseconds. When set to None, the
+        default, times and frequencies are treated as dimensionless quantities
+        that are scaled by the (undetermined) sampling time.
+
+    Examples
+    --------
+    >>> import thztools as thz
+    >>> thz.set_option("sampling_time", 0.05)
+    >>> thz.get_option("sampling_time")
+    0.05
+    >>> thz.set_option("sampling_time", None)
+    >>> thz.get_option("sampling_time")
     """
     setattr(options, key, value)
 
@@ -68,6 +87,24 @@ def get_option(key: str) -> Any:
     -------
     val : object
         Option value.
+
+    Notes
+    -----
+    Available options, with descriptions:
+
+    sampling_time : float | None
+        Global sampling time, normally in picoseconds. When set to None, the
+        default, times and frequencies are treated as dimensionless quantities
+        that are scaled by the (undetermined) sampling time.
+
+    Examples
+    --------
+    >>> import thztools as thz
+    >>> thz.set_option("sampling_time", 0.05)
+    >>> thz.get_option("sampling_time")
+    0.05
+    >>> thz.set_option("sampling_time", None)
+    >>> thz.get_option("sampling_time")
     """
     return getattr(options, key)
 
@@ -153,12 +190,11 @@ class NoiseModel:
     >>> import thztools as thz
     >>> from matplotlib import pyplot as plt
     >>> n, dt = 256, 0.05
-    >>> thz.set_option("sampling_time", dt)
-    >>> t = thz.timebase(n)
+    >>> t = thz.timebase(n, dt=dt)
     >>> mu = thz.wave(n)
     >>> alpha, beta, tau = 1e-4, 1e-2, 1e-3
     >>> noise_model = thz.NoiseModel(sigma_alpha=alpha, sigma_beta=beta,
-    ...  sigma_tau=tau)
+    ...  sigma_tau=tau, dt=dt)
     >>> noise_variance = noise_model.noise_var(mu)
 
     >>> _, axs = plt.subplots(2, 1, sharex=True, layout="constrained")
@@ -226,12 +262,11 @@ class NoiseModel:
         >>> import thztools as thz
         >>> from matplotlib import pyplot as plt
         >>> n, dt = 256, 0.05
-        >>> thz.set_option("sampling_time", dt)
-        >>> t = thz.timebase(n)
-        >>> mu = thz.wave(n)
+        >>> t = thz.timebase(n, dt=dt)
+        >>> mu = thz.wave(n, dt=dt)
         >>> alpha, beta, tau = 1e-4, 1e-2, 1e-3
         >>> noise_model = thz.NoiseModel(sigma_alpha=alpha, sigma_beta=beta,
-        ... sigma_tau=tau)
+        ... sigma_tau=tau, dt=dt)
         >>> noise_variance = noise_model.noise_var(mu)
 
         >>> _, axs = plt.subplots(2, 1, sharex=True, layout="constrained")
@@ -317,12 +352,11 @@ class NoiseModel:
         >>> import thztools as thz
         >>> from matplotlib import pyplot as plt
         >>> n, dt = 256, 0.05
-        >>> thz.set_option("sampling_time", dt)
-        >>> t = thz.timebase(n)
-        >>> mu = thz.wave(n)
+        >>> t = thz.timebase(n, dt=dt)
+        >>> mu = thz.wave(n, dt=dt)
         >>> alpha, beta, tau = 1e-4, 1e-2, 1e-3
         >>> noise_model = thz.NoiseModel(sigma_alpha=alpha, sigma_beta=beta,
-        ... sigma_tau=tau)
+        ... sigma_tau=tau, dt=dt)
         >>> noise_amplitude = noise_model.noise_amp(mu)
 
         >>> _, axs = plt.subplots(2, 1, sharex=True, layout="constrained")
@@ -391,12 +425,11 @@ class NoiseModel:
         >>> import thztools as thz
         >>> from matplotlib import pyplot as plt
         >>> n, dt = 256, 0.05
-        >>> thz.set_option("sampling_time", dt)
-        >>> t = thz.timebase(n)
-        >>> mu = thz.wave(n)
+        >>> t = thz.timebase(n, dt=dt)
+        >>> mu = thz.wave(n, dt=dt)
         >>> alpha, beta, tau = 1e-4, 1e-2, 1e-3
         >>> noise_model = thz.NoiseModel(sigma_alpha=alpha, sigma_beta=beta,
-        ... sigma_tau=tau)
+        ... sigma_tau=tau, dt=dt)
         >>> noise = noise_model.noise_sim(mu, seed=1234)
 
         >>> _, axs = plt.subplots(2, 1, sharex=True, layout="constrained")
@@ -514,9 +547,9 @@ def transfer(
     numpy_sign_convention : bool, optional
         Adopt NumPy sign convention for harmonic time dependence, e.g, express
         a harmonic function with frequency :math:`\omega` as
-        :math:`x(t) = a e^{i\omega t}`. Default is ``True``. Uses the
-        convention more common in physics, :math:`x(t) = a e^{-i\omega t}`,
-        when set to ``False``.
+        :math:`x(t) = a e^{i\omega t}`. Default is ``True``. When set to
+        ``False``, uses the convention more common in physics,
+        :math:`x(t) = a e^{-i\omega t}`.
     args : tuple, optional
         Extra arguments passed to the transfer function.
 
@@ -563,16 +596,16 @@ def transfer(
 
     set ``fft_sign=False``.
 
+    >>> import numpy as np
     >>> import thztools as thz
     >>> from matplotlib import pyplot as plt
     >>> n, dt = 256, 0.05
-    >>> thz.set_option("sampling_time", dt)
-    >>> t = thz.timebase(n)
-    >>> x = thz.wave(n)
+    >>> t = thz.timebase(n, dt=dt)
+    >>> x = thz.wave(n, dt=dt)
     >>> def shiftscale(_w, _a, _tau):
     ...     return _a * np.exp(-1j * _w * _tau)
     >>>
-    >>> y = thz.transfer(shiftscale, x,
+    >>> y = thz.transfer(shiftscale, x, dt=dt,
     ...                  numpy_sign_convention=True,
     ...                  args=(0.5, 1))
 
@@ -580,11 +613,9 @@ def transfer(
     >>>
     >>> ax.plot(t, x, label='x')
     >>> ax.plot(t, y, label='y')
-    >>>
     >>> ax.legend()
     >>> ax.set_xlabel('t (ps)')
     >>> ax.set_ylabel('Amplitude (arb. units)')
-    >>>
     >>> plt.show()
 
     >>> def shiftscale_phys(_w, _a, _tau):
@@ -596,12 +627,12 @@ def transfer(
 
     >>> _, ax = plt.subplots()
     >>>
-    >>> ax.plot(t, x, label='x');
-    >>> ax.plot(t, y_p, label='y');
+    >>> ax.plot(t, x, label='x')
+    >>> ax.plot(t, y_p, label='y')
     >>>
-    >>> ax.legend();
-    >>> ax.set_xlabel('t (ps)');
-    >>> ax.set_ylabel('Amplitude (arb. units)');
+    >>> ax.legend()
+    >>> ax.set_xlabel('t (ps)')
+    >>> ax.set_ylabel('Amplitude (arb. units)')
     >>>
     >>> plt.show()
     """
@@ -669,6 +700,7 @@ def timebase(
     assigning the sampling time.
 
     >>> import thztools as thz
+    >>> thz.set_option("sampling_time", None)
     >>> n, dt, t_init = 256, 0.05, 2.5
     >>> t_1 = thz.timebase(n)
     >>> t_2 = thz.timebase(n, dt=dt)
@@ -762,14 +794,13 @@ def wave(
     >>> import thztools as thz
     >>> from matplotlib import pyplot as plt
     >>> n, dt = 256, 0.05
-    >>> thz.set_option("sampling_time", dt)
-    >>> t = thz.timebase(n)
-    >>> mu = thz.wave(n)
+    >>> t = thz.timebase(n, dt=dt)
+    >>> mu = thz.wave(n, dt=dt)
 
     >>> _, ax = plt.subplots(layout="constrained")
-    >>> ax.plot(t, mu);
-    >>> ax.set_xlabel("t (ps)");
-    >>> ax.set_ylabel(r"$\mu/\mu_0$");
+    >>> ax.plot(t, mu)
+    >>> ax.set_xlabel("t (ps)")
+    >>> ax.set_ylabel(r"$\mu/\mu_0$")
     >>> plt.show()
     """
     dt = _assign_sampling_time(dt)
@@ -850,22 +881,22 @@ def scaleshift(
     rescale each copy by ``a = [1.0, 0.5, 0.25, 0.125]`` and shift it by
     ``eta = [0.0, 1.0, 2.0, 3.0]``.
 
+    >>> import numpy as np
     >>> import thztools as thz
     >>> from matplotlib import pyplot as plt
     >>> n, dt = 256, 0.05
-    >>> thz.set_option("sampling_time", dt)
-    >>> t = thz.timebase(n)
-    >>> mu = thz.wave(n)
+    >>> t = thz.timebase(n, dt=dt)
+    >>> mu = thz.wave(n, dt=dt)
     >>> m = 4
     >>> x = np.repeat(np.atleast_2d(mu), m, axis=0)
     >>> a = 0.5**np.arange(m)
     >>> eta = np.arange(m)
     >>> x_adj = thz.scaleshift(x, a=a, eta=eta, dt=dt)
 
-    >>> plt.plot(t, x_adj.T, label=[f"{k=}" for k in range(4)]);
-    >>> plt.legend();
-    >>> plt.xlabel("t (ps)");
-    >>> plt.ylabel(r"$x_{\mathrm{adj}, k}$");
+    >>> plt.plot(t, x_adj.T, label=[f"{k=}" for k in range(4)])
+    >>> plt.legend()
+    >>> plt.xlabel("t (ps)")
+    >>> plt.ylabel(r"$x_{\mathrm{adj}, k}$")
     >>> plt.show()
     """
     x = np.asarray(x, dtype=np.float64)
@@ -982,13 +1013,10 @@ def _costfuntls(
 
     dt = _assign_sampling_time(dt)
 
-    n = x.shape[-1]
+    psi = transfer(lambda omega: fun(theta, omega), mu, dt=dt)
+
     delta_norm = (x - mu) / sigma_x
-    w = 2 * np.pi * rfftfreq(n) / dt
-    h_f = fun(theta, w)
-
-    eps_norm = (y - irfft(rfft(mu) * h_f, n=n)) / sigma_y
-
+    eps_norm = (y - psi) / sigma_y
     res = np.concatenate((delta_norm, eps_norm))
 
     return res
@@ -1343,32 +1371,33 @@ def noisefit(
     with ``n = 256`` time points each, then verify that the fit results are
     consistent with the true noise parameters.
 
+    >>> import numpy as np
     >>> import thztools as thz
     >>> from matplotlib import pyplot as plt
     >>> rng = np.random.default_rng(0)
     >>> n, m, dt = 256, 50, 0.05
-    >>> thz.set_option("sampling_time", dt)
-    >>> t = thz.timebase(n)
-    >>> mu = thz.wave(n)
-    >>> alpha, beta, tau = 1e-5, 1e-2, 1e-3
+    >>> t = thz.timebase(n, dt=dt)
+    >>> mu = thz.wave(n, dt=dt)
+    >>> alpha, beta, tau = 1e-4, 1e-2, 1e-3
     >>> noise_model = thz.NoiseModel(sigma_alpha=alpha, sigma_beta=beta,
-    ...  sigma_tau=tau)
+    ...  sigma_tau=tau, dt=dt)
     >>> a = 1.0 + 1e-2 * np.concatenate(([0.0],
     ...                                 rng.standard_normal(m - 1)))
     >>> eta = 1e-3 * np.concatenate(([0.0], rng.standard_normal(m - 1)))
-    >>> z = thz.scaleshift(np.repeat(np.atleast_2d(mu), m, axis=0),
+    >>> z = thz.scaleshift(np.repeat(np.atleast_2d(mu), m, axis=0), dt=dt,
     ...                    a=a, eta=eta).T  # Orient the array columnwise
     >>> x = z + noise_model.noise_sim(z, axis=1, seed=1234)
-    >>> noise_res = thz.noisefit(x)
+    >>> noise_res = thz.noisefit(x, sigma_alpha0=alpha, sigma_beta0=beta,
+    ...  sigma_tau0=tau, dt=dt)
     >>> noise_res.noise_model  #doctest: +NORMALIZE_WHITESPACE
-    NoiseModel(sigma_alpha=1.000...e-05, sigma_beta=0.009609...,
-    sigma_tau=0.0009243..., dt=0.05)
+    NoiseModel(sigma_alpha=9.9...e-05, sigma_beta=0.0096...,
+    sigma_tau=4.27...e-06, dt=0.05)
 
     >>> plt.plot(t, np.std(thz.scaleshift(x, a=1 / noise_res.a,
     ... eta=-noise_res.eta, axis=0), axis=1), "-",
     ... label="Data")
     >>> plt.plot(t, noise_res.noise_model.noise_amp(noise_res.mu)
-    ...  * np.sqrt(m / (m + 1)), "--", label="Fit")
+    ...  * np.sqrt(m / (m - 1)), "--", label="Fit")
     >>> plt.legend()
     >>> plt.xlabel("t (ps)")
     >>> plt.ylabel(r"$\sigma(t)$")
@@ -1693,7 +1722,7 @@ def _parse_noisefit_output(
     if fix_sigma_tau:
         tau = sigma_tau0
     else:
-        tau = np.exp(x_out[0] / 2) * scale_sigma_tau * dt
+        tau = np.exp(x_out[0] / 2) * scale_sigma_tau
         x_out = x_out[1:]
     # noinspection PyArgumentList
     noise_model = NoiseModel(alpha, beta, tau, dt=dt)
@@ -1801,85 +1830,126 @@ def _parse_noisefit_output(
     )
 
 
+@dataclass
+class FitResult:
+    r"""
+    Dataclass for the output of :func:`fit`.
+
+    Parameters
+    ----------
+    p_opt : array_like
+        Optimal fit parameters.
+    p_var : array_like
+        Variance of p_opt.
+    mu_opt : array_like
+        Optimal underlying waveform.
+    mu_var : array_like
+        Variance of mu_opt.
+    resnorm : float
+        The value of chi-squared.
+    delta : array_like
+        Residuals of the input waveform ``x``, defined as ``x - mu_opt``.
+    epsilon : array_like
+        Residuals of the output waveform ``y``, defined as ``y - psi_opt``,
+        where ``psi_opt = thztools.transfer(tfun_opt, mu, dt=dt)`` and
+        ``tfun_opt`` is the parameterized transfer function evaluated at
+        ``p_opt``.
+    r_tls : array_like
+        Total least-squares residuals.
+    success : bool
+        True if one of the convergence criteria is satisfied.
+
+    See Also
+    --------
+    fit : Fit a transfer function to time-domain data.
+    """
+    p_opt: NDArray[np.float64]
+    p_var: NDArray[np.float64]
+    mu_opt: NDArray[np.float64]
+    mu_var: NDArray[np.float64]
+    resnorm: float
+    delta: NDArray[np.float64]
+    epsilon: NDArray[np.float64]
+    r_tls: NDArray[np.float64]
+    success: bool
+
+
 def fit(
     fun: Callable,
+    xdata: ArrayLike,
+    ydata: ArrayLike,
     p0: ArrayLike,
-    x: ArrayLike,
-    y: ArrayLike,
+    noise_parms: ArrayLike | None = None,
     *,
     dt: float | None = None,
-    sigma_parms: ArrayLike = (1.0, 0.0, 0.0),
-    f_bounds: ArrayLike = (0.0, np.inf),
+    numpy_sign_convention: bool = True,
+    f_bounds: ArrayLike | None = None,
     p_bounds: ArrayLike | None = None,
     jac: Callable | None = None,
-    args: ArrayLike = (),
-    kwargs: dict | None = None,
-) -> dict:
+) -> FitResult:
     r"""
     Fit a transfer function to time-domain data.
 
-    Computes the noise on the input ``x`` and output ``y`` time series using
-    a given ``NoiseModel``. Uses the total residuals generated by
-    ``_costfuntls`` to fit the input and output to the transfer function.
+    Determines the total least-squares fit to ``xdata`` and ``ydata``, given
+    the parameterized transfer function relationship ``fun`` and noise model
+    parameters ``noise_parms``.
 
     Parameters
     ----------
     fun : callable
-        Transfer function.
-
-            ``fun(p, w, *args, **kwargs) -> ndarray``
-
-        Assumes the :math:`+i\omega t` convention for harmonic time dependence.
-    p0 : array_like
-        Initial guess for ``p``.
-    x : array_like
+        Transfer function with signature ``fun(omega, *p, *args, **kwargs)``
+        that returns an ``ndarray``. Assumes the :math:`+i\omega t` convention
+        for harmonic time dependence when ``numpy_sign_convention`` is
+        ``True``, the default.
+    xdata : array_like
         Measured input signal.
-    y : array_like
+    ydata : array_like
         Measured output signal.
+    p0 : array_like
+        Initial guess for the parameters ``p``.
+    noise_parms : array_like or None, optional
+        Noise parameters ``(sigma_alpha, sigma_beta, sigma_tau)`` used to
+        define the :class:``NoiseModel`` for the fit. Default is ``None``,
+        which sets ``noise_parms`` to ``(1.0, 0.0, 0.0)``.
     dt : float or None, optional
-        Sampling time, normally in picoseconds. Default is None, which sets
-        the sampling time to ``thztools.options.sampling_time``. If both
-        ``dt`` and ``thztools.options.sampling_time`` are ``None``, the
-        sampling time is set to ``1.0``. In this case, the angular frequency
-        ``omega`` must be given in units of radians per sampling time, and any
-        parameters in ``args`` must be expressed with the sampling time as the
-        unit of time.
-    sigma_parms : None or array_like, optional
-        Noise parameters with size (3,), expressed as standard deviation
-        amplitudes.
+        Sampling time, normally in picoseconds. Default is ``None``, which sets
+        ``dt`` to ``thztools.options.sampling_time``. If both ``dt`` and
+        ``thztools.options.sampling_time`` are ``None``, ``dt`` is set to
+        ``1.0``. In this case, the angular frequency ``omega`` must be given in
+        units of radians per sampling time, and any parameters in ``args`` must
+        be expressed with ``dt`` as the unit of time.
+    numpy_sign_convention : bool, optional
+        Adopt NumPy sign convention for harmonic time dependence, e.g, express
+        a harmonic function with frequency :math:`\omega` as
+        :math:`x(t) = a e^{i\omega t}`. Default is ``True``. When set to
+        ``False``, uses the convention more common in physics,
+        :math:`x(t) = a e^{-i\omega t}`.
     f_bounds : array_like, optional
-        Frequency bounds.
+        Frequency bounds. Default is ``None``, which sets ``f_bounds`` to
+        ``(0.0, np.inf)``.
     p_bounds : None, 2-tuple of array_like, or Bounds, optional
-        Lower and upper bounds on fit parameter(s).
+        Lower and upper bounds on fit parameter(s). Default is ``None``, which
+        sets all parameter bounds to ``(-np.inf, np.inf)``.
     jac : None or callable, optional
-        Method of calculating derivative of the output signal residuals
-        with respect to the fit parameter(s), theta.
-    args : tuple, optional
-        Additional arguments passed to ``fun`` and ``jac``.
-    kwargs : dict, optional
-        Additional keyword arguments passed to ``fun`` and ``jac``.
+        Jacobian of the residuals with respect to the fit parameters, with
+        signature ``jac(p, w, *args, **kwargs)``. Default is ``None``, which
+        uses :func:`scipy.optimize.approx_fprime`.
 
     Returns
     -------
-    p : dict
-        Output parameter dictionary containing:
+    res : FitResult
+        Fit result represented as a :class:`FitResult` object. Important
+        attributes are: ``p_opt``, the optimal fit parameter values; ``p_cov``,
+        the parameter covariance matrix estimated from fit; ``resnorm``,
+        the value of the total least-squares cost function for the fit;
+        ``r_tls``, and ``success``, which is `True` when the fit converges. See
+        the *Notes* section below and the :class:`FitResult` documentation for
+        more details and for a description of other attributes.
 
-            p_opt : array_like
-                Optimal fit parameters.
-            p_cov : array_like
-                Variance of p_opt.
-            mu_opt : array_like
-                Optimal underlying waveform.
-            mu_var : array_like
-                Variance of mu_opt.
-            resnorm : float
-                The value of chi-squared.
-            delta : array_like
-                Residuals of the input waveform ``x``.
-            epsilon : array_like
-                Resiuals of the output waveform ``y``.
-            success : bool
-                True if one of the convergence criteria is satisfied.
+    Raises
+    ------
+    ValueError
+        If ``noise_parms`` does not have 3 elements.
 
     Warns
     -----
@@ -1888,18 +1958,106 @@ def fit(
         are both not ``None`` and are set to different ``float`` values, the
         function will set the sampling time to ``dt`` and raise a
         :class:`UserWarning`.
+
+    See Also
+    --------
+    NoiseModel : Noise model class.
+
+    Notes
+    -----
+    This function computes the maximum-likelihood estimate for the parameters
+    :math:`\boldsymbol{\theta}` in the transfer function model
+    :math:`H(\omega; \boldsymbol{\theta})` by minimizing
+    the total least-squares cost function
+
+    .. math:: Q_\text{TLS}(\boldsymbol{\theta}) \
+        = \sum_{k=0}^{N-1}\left[ \
+        \frac{(x_k - \mu_k)^2}{\sigma_{\mathbf{x},k}^2} \
+        + \frac{(y_k - \psi_k)^2}{\sigma_{\mathbf{y},k}^2} \
+        \right],
+
+    where :math:`\mathbf{x}` is the input signal array, :math:`\mathbf{y}`
+    is the output signal array, and :math:`\boldsymbol{\sigma}_{\mathbf{y}}`,
+    :math:`\boldsymbol{\sigma}_{\mathbf{y}}` are their associated noise
+    amplitudes. The arrays :math:`\boldsymbol{\mu}` and
+    :math:`\boldsymbol{\psi}` are the ideal input and output signal arrays,
+    respectively, which are related by the constraint
+
+    .. math:: [\mathcal{F}\boldsymbol{\psi}]_k = \
+        [H(\omega_k; \boldsymbol{\theta})] \
+        [\mathcal{F}\boldsymbol{\mu}]_k
+
+    at all discrete Fourier transform frequencies :math:`\omega_k` within
+    the frequency bounds.
+
+    The optimization step uses :func:`scipy.optimize.least_squares` with ``p``
+    and ``mu`` as free parameters. It minimizes the Euclidean norm of the
+    residual vector ``np.concat((delta / sigma_x, epsilon / sigma_y)``, where
+    ``delta = xdata - mu``, ``epsilon = ydata - psi``,
+    ``sigma_x = NoiseModel(*noise_parms, dt=dt).noise_amp(xdata)``,
+    ``sigma_y = NoiseModel(*noise_parms, dt=dt).noise_amp(ydata)``, and
+    ``psi = thz.transfer(fun(omega, *p, *args, **kwargs), mu, dt=dt)``.
+
+    References
+    ----------
+    .. [1] Laleh Mohtashemi, Paul Westlund, Derek G. Sahota, Graham B. Lea,
+        Ian Bushfield, Payam Mousavi, and J. Steven Dodge, "Maximum-
+        likelihood parameter estimation in terahertz time-domain
+        spectroscopy," Opt. Express **29**, 4912-4926 (2021),
+        `<https://doi.org/10.1364/OE.417724>`_.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> import thztools as thz
+    >>> from matplotlib import pyplot as plt
+    >>> n, dt = 256, 0.05
+    >>> t = thz.timebase(n, dt=dt)
+    >>> mu = thz.wave(n, dt=dt)
+    >>> alpha, beta, tau = 1e-4, 1e-2, 1e-3
+    >>> noise_model = thz.NoiseModel(sigma_alpha=alpha, sigma_beta=beta,
+    ...  sigma_tau=tau, dt=dt)
+
+    >>> def tfun(p, w):
+    ...    return p[0] * np.exp(1j * p[1] * w)
+    >>>
+    >>> p0 = (0.5, 1.0)
+    >>> psi = thz.transfer(lambda omega: tfun(p0, omega), mu, dt=dt)
+    >>> xdata = mu + noise_model.noise_sim(mu, seed=0)
+    >>> ydata = psi + noise_model.noise_sim(psi, seed=1)
+    >>> result = thz.fit(tfun, xdata, ydata, p0, (alpha, beta, tau), dt=dt)
+    >>> result.p_opt
+    array([0.499..., 1.000...])
+    >>> result.resnorm
+    198.300...
+
+    >>> _, ax = plt.subplots()
+    >>> ax.plot(t, result.r_tls, '.')
+    >>> ax.set_xlabel("t (ps)")
+    >>> ax.set_ylabel(r"$r_\mathrm{TLS}$")
+    >>> plt.show()
     """
     fit_method = "trf"
 
     p0 = np.asarray(p0, dtype=np.float64)
-    x = np.asarray(x, dtype=np.float64)
-    y = np.asarray(y, dtype=np.float64)
-    sigma_parms = np.asarray(sigma_parms, dtype=np.float64)
+    xdata = np.asarray(xdata, dtype=np.float64)
+    ydata = np.asarray(ydata, dtype=np.float64)
+    if noise_parms is None:
+        noise_parms = np.asarray((1.0, 0.0, 0.0), dtype=np.float64)
+    else:
+        noise_parms = np.asarray(noise_parms, dtype=np.float64)
+
+    if noise_parms.size != NUM_NOISE_PARAMETERS:
+        msg = f"sigma_parms must be a tuple of length {NUM_NOISE_PARAMETERS:d}"
+        raise ValueError(msg)
 
     dt = _assign_sampling_time(dt)
 
-    n = y.shape[-1]
+    n = ydata.shape[-1]
     n_p = len(p0)
+
+    if f_bounds is None:
+        f_bounds = np.array((0.0, np.inf), dtype=np.float64)
 
     if p_bounds is None:
         p_bounds = (-np.inf, np.inf)
@@ -1915,9 +2073,6 @@ def fit(
             msg = "`bounds` must contain 2 elements."
             raise ValueError(msg)
 
-    if kwargs is None:
-        kwargs = {}
-
     w = 2 * np.pi * rfftfreq(n, dt)
     w_bounds = 2 * np.pi * np.asarray(f_bounds, dtype=np.float64)
     w_below_idx = w <= w_bounds[0]
@@ -1925,45 +2080,50 @@ def fit(
     w_in_idx = np.invert(w_below_idx) * np.invert(w_above_idx)
     n_f = len(w)
 
-    alpha, beta, tau = sigma_parms
+    alpha, beta, tau = noise_parms
     # noinspection PyArgumentList
     noise_model = NoiseModel(alpha, beta, tau, dt=dt)
-    sigma_x = noise_model.noise_amp(x)
-    sigma_y = noise_model.noise_amp(y)
+    sigma_x = noise_model.noise_amp(xdata)
+    sigma_y = noise_model.noise_amp(ydata)
     p0_est = np.concatenate((p0, np.zeros(n)))
 
     def etfe(_x, _y):
-        return rfft(_y) / rfft(_x)
+        h = rfft(_y) / rfft(_x)
+        if numpy_sign_convention:
+            h = np.conj(h)
+        return h
 
     def function(_theta, _w):
-        _h = etfe(x, y)
+        _h = etfe(xdata, ydata)
         _w_in = _w[w_in_idx]
-        return np.concatenate(
-            (
-                _h[w_below_idx],
-                fun(_theta, _w_in, *args, **kwargs),
-                _h[w_above_idx],
-            )
-        )
+        h_lo = _h[w_below_idx]
+        h_in = fun(_theta, _w_in)
+        h_hi = _h[w_above_idx]
+        if not numpy_sign_convention:
+            h_in = np.conj(h_in)
+        return np.concatenate((h_lo, h_in, h_hi))
 
     def function_flat(_x):
         _tf = function(_x, w)
         return np.concatenate((np.real(_tf), np.imag(_tf)))
 
     def td_fun(_p, _x):
-        _h = irfft(rfft(_x) * function(_p, w), n=n)
+        if numpy_sign_convention:
+            _h = irfft(rfft(_x) * function(_p, w), n=n)
+        else:
+            _h = irfft(rfft(_x) * np.conj(function(_p, w)), n=n)
         return _h
 
     def jacobian(_p):
         if jac is None:
-            _tf_prime = fprime(_p, function_flat)
+            _tf_prime = approx_fprime(_p, function_flat)
             return _tf_prime[0:n_f] + 1j * _tf_prime[n_f:]
         else:
-            return jac(_p, w, *args, **kwargs)
+            return jac(_p, w)
 
     def jac_fun(_x):
         p_est = _x[:n_p]
-        mu_est = x[:] - _x[n_p:]
+        mu_est = xdata[:] - _x[n_p:]
         jac_tl = np.zeros((n, n_p))
         jac_tr = np.diag(1 / sigma_x)
         jac_bl = -(
@@ -1980,9 +2140,9 @@ def fit(
         lambda _p: _costfuntls(
             function,
             _p[:n_p],
-            x[:] - _p[n_p:],
-            x[:],
-            y[:],
+            xdata[:] - _p[n_p:],
+            xdata[:],
+            ydata[:],
             sigma_x,
             sigma_y,
             dt,
@@ -1995,17 +2155,35 @@ def fit(
     )
 
     # Parse output
-    p = {}
-    p["p_opt"] = result.x[:n_p]
-    p["mu_opt"] = x - result.x[n_p:]
     _, s, vt = la.svd(result.jac, full_matrices=False)
     threshold = np.finfo(float).eps * max(result.jac.shape) * s[0]
     s = s[s > threshold]
     vt = vt[: s.size]
-    p["p_var"] = np.diag(np.dot(vt.T / s**2, vt))[:n_p]
-    p["mu_var"] = np.diag(np.dot(vt.T / s**2, vt))[n_p:]
-    p["resnorm"] = 2 * result.cost
-    p["delta"] = x - p["mu_opt"]
-    p["epsilon"] = y - irfft(rfft(p["mu_opt"]) * function(p["p_opt"], w), n=n)
-    p["success"] = result.success
-    return p
+
+    p_opt = result.x[:n_p]
+    p_var = np.diag(np.dot(vt.T / s**2, vt))[n_p:]
+    delta = result.x[n_p:]
+    mu_opt = xdata - delta
+    mu_var = np.diag(np.dot(vt.T / s**2, vt))[n_p:]
+    resnorm = 2 * result.cost
+    psi_opt = transfer(lambda _w: function(p_opt, _w), mu_opt, dt=dt)
+    epsilon = ydata - psi_opt
+
+    v_y = np.diag(sigma_y**2)
+    h_sigma_x = transfer(lambda _w: function(p_opt, _w), sigma_x, dt=dt)
+    u_x = h_sigma_x[:, np.newaxis] @ h_sigma_x[np.newaxis, :]
+    h_delta = transfer(lambda _w: function(p_opt, _w), delta, dt=dt)
+    r_tls = sqrtm(np.linalg.inv(v_y + u_x)) @ (epsilon - h_delta)
+
+    res = FitResult(
+        p_opt=p_opt,
+        p_var=p_var,
+        mu_opt=mu_opt,
+        mu_var=mu_var,
+        resnorm=resnorm,
+        delta=delta,
+        epsilon=epsilon,
+        r_tls=r_tls,
+        success=result.success,
+    )
+    return res